extern cleanup: normalize externs of ~130 more functions to true signatures across 77 files (byte-identical, full -k0 verified)

+++ src/main/dll/dll_0041_warpstoneui.c
@@ -21,7 +21,7 @@ extern void gameTextSetColor(u8 r, u8 g, u8 b, u8 a);
 extern void gameTextShow(int a);
 extern void gameTextFn_80016810(int a, int b, int c);
 extern void drawTexture(void* tex, f32 x, f32 y, int alpha, int p5);
-extern void CMenu_SetFadeCounter(int v);
+extern void CMenu_SetFadeCounter(s16 v);
 
 extern u8 warpstoneUIState;
 extern void* lbl_803DD984;
